Rebase from master to RAC5
