+++ ReactiveCocoa/Cocoa/FoundationExtensions.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import ReactiveSwift
 import enum Result.NoError
 
 extension NSNotificationCenter {
