+++ ReactiveCocoaTests/Bridging/ObjectiveCBridgingSpec.swift
@@ -9,7 +9,7 @@
 import Result
 import Nimble
 import Quick
-import ReactiveCocoa
+import ReactiveSwift
 import XCTest
 
 class ObjectiveCBridgingSpec: QuickSpec {
