+++ ReactiveCocoaTests/Swift/PropertySpec.swift
@@ -9,7 +9,7 @@
 import Result
 import Nimble
 import Quick
-import ReactiveCocoa
+import ReactiveSwift
 
 private let initialPropertyValue = "InitialValue"
 private let subsequentPropertyValue = "SubsequentValue"
@@ -513,777 +513,17 @@ class PropertySpec: QuickSpec {
 			}
 		}
 
-		describe("PropertyType") {
-			describe("map") {
-				it("should transform the current value and all subsequent values") {
-					let property = MutableProperty(1)
-					let mappedProperty = property
-						.map { $0 + 1 }
-					expect(mappedProperty.value) == 2
-
-					property.value = 2
-					expect(mappedProperty.value) == 3
-				}
-			}
-
-			describe("combineLatest") {
-				var property: MutableProperty<String>!
-				var otherProperty: MutableProperty<String>!
-
-				beforeEach {
-					property = MutableProperty(initialPropertyValue)
-					otherProperty = MutableProperty(initialOtherPropertyValue)
-				}
-
-				it("should forward the latest values from both inputs") {
-					let combinedProperty = property.combineLatestWith(otherProperty)
-					var latest: (String, String)?
-					combinedProperty.signal.observeNext { latest = $0 }
-
-					property.value = subsequentPropertyValue
-					expect(latest?.0) == subsequentPropertyValue
-					expect(latest?.1) == initialOtherPropertyValue
-
-					// is there a better way to test tuples?
-					otherProperty.value = subsequentOtherPropertyValue
-					expect(latest?.0) == subsequentPropertyValue
-					expect(latest?.1) == subsequentOtherPropertyValue
-
-					property.value = finalPropertyValue
-					expect(latest?.0) == finalPropertyValue
-					expect(latest?.1) == subsequentOtherPropertyValue
-				}
-
-				it("should complete when the source properties are deinitialized") {
-					var completed = false
-
-					var combinedProperty = Optional(property.combineLatestWith(otherProperty))
-					combinedProperty!.signal.observeCompleted { completed = true }
-
-					combinedProperty = nil
-					expect(completed) == false
-
-					property = nil
-					expect(completed) == false
-
-					otherProperty = nil
-					expect(completed) == true
-				}
-			}
-
-			describe("zip") {
-				var property: MutableProperty<String>!
-				var otherProperty: MutableProperty<String>!
-
-				beforeEach {
-					property = MutableProperty(initialPropertyValue)
-					otherProperty = MutableProperty(initialOtherPropertyValue)
-				}
-
-				it("should combine pairs") {
-					var result: [String] = []
-
-					let zippedProperty = property.zipWith(otherProperty)
-					zippedProperty.producer.startWithNext { (left, right) in result.append("\(left)\(right)") }
-
-					let firstResult = [ "\(initialPropertyValue)\(initialOtherPropertyValue)" ]
-					let secondResult = firstResult + [ "\(subsequentPropertyValue)\(subsequentOtherPropertyValue)" ]
-					let thirdResult = secondResult + [ "\(finalPropertyValue)\(finalOtherPropertyValue)" ]
-					let finalResult = thirdResult + [ "\(initialPropertyValue)\(initialOtherPropertyValue)" ]
-
-					expect(result) == firstResult
-
-					property.value = subsequentPropertyValue
-					expect(result) == firstResult
-
-					otherProperty.value = subsequentOtherPropertyValue
-					expect(result) == secondResult
-
-					property.value = finalPropertyValue
-					otherProperty.value = finalOtherPropertyValue
-					expect(result) == thirdResult
-
-					property.value = initialPropertyValue
-					expect(result) == thirdResult
-
-					property.value = subsequentPropertyValue
-					expect(result) == thirdResult
-
-					otherProperty.value = initialOtherPropertyValue
-					expect(result) == finalResult
-				}
-
-				it("should complete its producer only when the source properties are deinitialized") {
-					var result: [String] = []
-					var completed = false
-
-					var zippedProperty = Optional(property.zipWith(otherProperty))
-					zippedProperty!.producer.start { event in
-						switch event {
-						case let .Next(left, right):
-							result.append("\(left)\(right)")
-						case .Completed:
-							completed = true
-						default:
-							break
-						}
-					}
-
-					expect(completed) == false
-					expect(result) == [ "\(initialPropertyValue)\(initialOtherPropertyValue)" ]
-
-					property.value = subsequentPropertyValue
-					expect(result) == [ "\(initialPropertyValue)\(initialOtherPropertyValue)" ]
-
-					zippedProperty = nil
-					expect(completed) == false
-
-					property = nil
-					otherProperty = nil
-					expect(completed) == true
-				}
-			}
-
-			describe("unary operators") {
-				var property: MutableProperty<String>!
-
-				beforeEach {
-					property = MutableProperty(initialPropertyValue)
-				}
-
-				describe("combinePrevious") {
-					it("should pack the current value and the previous value a tuple") {
-						let transformedProperty = property.combinePrevious(initialPropertyValue)
-
-						expect(transformedProperty.value.0) == initialPropertyValue
-						expect(transformedProperty.value.1) == initialPropertyValue
-
-						property.value = subsequentPropertyValue
-
-						expect(transformedProperty.value.0) == initialPropertyValue
-						expect(transformedProperty.value.1) == subsequentPropertyValue
-
-						property.value = finalPropertyValue
-
-						expect(transformedProperty.value.0) == subsequentPropertyValue
-						expect(transformedProperty.value.1) == finalPropertyValue
-					}
-
-					it("should complete its producer only when the source property is deinitialized") {
-						var result: (String, String)?
-						var completed = false
-
-						var transformedProperty = Optional(property.combinePrevious(initialPropertyValue))
-						transformedProperty!.producer.start { event in
-							switch event {
-							case let .Next(tuple):
-								result = tuple
-							case .Completed:
-								completed = true
-							default:
-								break
-							}
-						}
-
-						expect(result?.0) == initialPropertyValue
-						expect(result?.1) == initialPropertyValue
-
-						property.value = subsequentPropertyValue
-
-						expect(result?.0) == initialPropertyValue
-						expect(result?.1) == subsequentPropertyValue
-
-						transformedProperty = nil
-						expect(completed) == false
-
-						property = nil
-						expect(completed) == true
-					}
-				}
-
-				describe("skipRepeats") {
-					it("should not emit events for subsequent equatable values that are the same as the current value") {
-						let transformedProperty = property.skipRepeats()
-
-						var counter = 0
-						transformedProperty.signal.observeNext { _ in
-							counter += 1
-						}
-
-						property.value = initialPropertyValue
-						property.value = initialPropertyValue
-						property.value = initialPropertyValue
-
-						expect(counter) == 0
-
-						property.value = subsequentPropertyValue
-						property.value = subsequentPropertyValue
-						property.value = subsequentPropertyValue
-
-						expect(counter) == 1
-
-						property.value = finalPropertyValue
-						property.value = initialPropertyValue
-						property.value = subsequentPropertyValue
-
-						expect(counter) == 4
-					}
-
-					it("should not emit events for subsequent values that are regarded as the same as the current value by the supplied closure") {
-						var counter = 0
-						let transformedProperty = property.skipRepeats { _, newValue in newValue == initialPropertyValue }
-
-						transformedProperty.signal.observeNext { _ in
-							counter += 1
-						}
-
-						property.value = initialPropertyValue
-						expect(counter) == 0
-
-						property.value = subsequentPropertyValue
-						expect(counter) == 1
-
-						property.value = finalPropertyValue
-						expect(counter) == 2
-
-						property.value = initialPropertyValue
-						expect(counter) == 2
-					}
-
-					it("should complete its producer only when the source property is deinitialized") {
-						var counter = 0
-						var completed = false
-
-						var transformedProperty = Optional(property.skipRepeats())
-						transformedProperty!.producer.start { event in
-							switch event {
-							case .Next:
-								counter += 1
-							case .Completed:
-								completed = true
-							default:
-								break
-							}
-						}
-
-						expect(counter) == 1
-
-						property.value = initialPropertyValue
-						expect(counter) == 1
-
-						transformedProperty = nil
-						expect(completed) == false
-
-						property = nil
-						expect(completed) == true
-					}
-				}
-
-				describe("uniqueValues") {
-					it("should emit hashable values that have not been emited before") {
-						let transformedProperty = property.uniqueValues()
-
-						var counter = 0
-						transformedProperty.signal.observeNext { _ in
-							counter += 1
-						}
-
-						property.value = initialPropertyValue
-						expect(counter) == 0
-
-						property.value = subsequentPropertyValue
-						property.value = subsequentPropertyValue
-
-						expect(counter) == 1
-
-						property.value = finalPropertyValue
-						property.value = initialPropertyValue
-						property.value = subsequentPropertyValue
-
-						expect(counter) == 2
-					}
-
-					it("should emit only the values of which the computed identity have not been captured before") {
-						let transformedProperty = property.uniqueValues { _ in 0 }
-
-						var counter = 0
-						transformedProperty.signal.observeNext { _ in
-							counter += 1
-						}
-
-						property.value = initialPropertyValue
-						property.value = subsequentPropertyValue
-						property.value = finalPropertyValue
-						expect(counter) == 0
-					}
-
-					it("should complete its producer only when the source property is deinitialized") {
-						var counter = 0
-						var completed = false
-
-						var transformedProperty = Optional(property.uniqueValues())
-						transformedProperty!.producer.start { event in
-							switch event {
-							case .Next:
-								counter += 1
-							case .Completed:
-								completed = true
-							default:
-								break
-							}
-						}
-
-						expect(counter) == 1
-
-						property.value = initialPropertyValue
-						expect(counter) == 1
-
-						transformedProperty = nil
-						expect(completed) == false
-
-						property = nil
-						expect(completed) == true
-					}
-				}
-			}
-
-			describe("flattening") {
-				describe("flatten") {
-					describe("FlattenStrategy.Concat") {
-						it("should concatenate the values as the inner property is replaced and deinitialized") {
-							var firstProperty = Optional(MutableProperty(0))
-							var secondProperty = Optional(MutableProperty(10))
-							var thirdProperty = Optional(MutableProperty(20))
-
-							var outerProperty = Optional(MutableProperty(firstProperty!))
-
-							var receivedValues: [Int] = []
-							var errored = false
-							var completed = false
-
-							var flattenedProperty = Optional(outerProperty!.flatten(.Concat))
-
-							flattenedProperty!.producer.start { event in
-								switch event {
-								case let .Next(value):
-									receivedValues.append(value)
-								case .Completed:
-									completed = true
-								case .Failed:
-									errored = true
-								case .Interrupted:
-									break
-								}
-							}
-
-							expect(receivedValues) == [ 0 ]
-
-							outerProperty!.value = secondProperty!
-							secondProperty!.value = 11
-							outerProperty!.value = thirdProperty!
-							thirdProperty!.value = 21
-
-							expect(receivedValues) == [ 0 ]
-							expect(completed) == false
-
-							secondProperty!.value = 12
-							thirdProperty!.value = 22
-
-							expect(receivedValues) == [ 0 ]
-							expect(completed) == false
-
-							firstProperty = nil
-
-							expect(receivedValues) == [ 0, 12 ]
-							expect(completed) == false
-
-							secondProperty = nil
-
-							expect(receivedValues) == [ 0, 12, 22 ]
-							expect(completed) == false
-
-							outerProperty = nil
-							expect(completed) == false
-
-							thirdProperty = nil
-							expect(completed) == false
-
-							flattenedProperty = nil
-							expect(completed) == true
-							expect(errored) == false
-						}
-					}
-
-					describe("FlattenStrategy.Merge") {
-						it("should merge the values of all inner properties") {
-							var firstProperty = Optional(MutableProperty(0))
-							var secondProperty = Optional(MutableProperty(10))
-							var thirdProperty = Optional(MutableProperty(20))
-
-							var outerProperty = Optional(MutableProperty(firstProperty!))
-
-							var receivedValues: [Int] = []
-							var errored = false
-							var completed = false
-
-							var flattenedProperty = Optional(outerProperty!.flatten(.Merge))
-
-							flattenedProperty!.producer.start { event in
-								switch event {
-								case let .Next(value):
-									receivedValues.append(value)
-								case .Completed:
-									completed = true
-								case .Failed:
-									errored = true
-								case .Interrupted:
-									break
-								}
-							}
-
-							expect(receivedValues) == [ 0 ]
-
-							outerProperty!.value = secondProperty!
-							secondProperty!.value = 11
-							outerProperty!.value = thirdProperty!
-							thirdProperty!.value = 21
-
-							expect(receivedValues) == [ 0, 10, 11, 20, 21 ]
-							expect(completed) == false
-
-							secondProperty!.value = 12
-							thirdProperty!.value = 22
-
-							expect(receivedValues) == [ 0, 10, 11, 20, 21, 12, 22 ]
-							expect(completed) == false
-
-							firstProperty = nil
-
-							expect(receivedValues) == [ 0, 10, 11, 20, 21, 12, 22 ]
-							expect(completed) == false
-
-							secondProperty = nil
-
-							expect(receivedValues) == [ 0, 10, 11, 20, 21, 12, 22 ]
-							expect(completed) == false
-
-							outerProperty = nil
-							expect(completed) == false
-
-							thirdProperty = nil
-							expect(completed) == false
-
-							flattenedProperty = nil
-							expect(completed) == true
-							expect(errored) == false
-						}
-					}
-
-					describe("FlattenStrategy.Latest") {
-						it("should forward values from the latest inner property") {
-							let firstProperty = Optional(MutableProperty(0))
-							var secondProperty = Optional(MutableProperty(10))
-							var thirdProperty = Optional(MutableProperty(20))
-
-							var outerProperty = Optional(MutableProperty(firstProperty!))
-
-							var receivedValues: [Int] = []
-							var errored = false
-							var completed = false
-
-							outerProperty!.flatten(.Latest).producer.start { event in
-								switch event {
-								case let .Next(value):
-									receivedValues.append(value)
-								case .Completed:
-									completed = true
-								case .Failed:
-									errored = true
-								case .Interrupted:
-									break
-								}
-							}
-
-							expect(receivedValues) == [ 0 ]
-
-							outerProperty!.value = secondProperty!
-							secondProperty!.value = 11
-							outerProperty!.value = thirdProperty!
-							thirdProperty!.value = 21
-
-							expect(receivedValues) == [ 0, 10, 11, 20, 21 ]
-							expect(errored) == false
-							expect(completed) == false
-
-							secondProperty!.value = 12
-							secondProperty = nil
-							thirdProperty!.value = 22
-							thirdProperty = nil
-
-							expect(receivedValues) == [ 0, 10, 11, 20, 21, 22 ]
-							expect(errored) == false
-							expect(completed) == false
-
-							outerProperty = nil
-							expect(errored) == false
-							expect(completed) == true
-						}
-
-						it("should release the old properties when switched or deallocated") {
-							var firstProperty = Optional(MutableProperty(0))
-							var secondProperty = Optional(MutableProperty(10))
-							var thirdProperty = Optional(MutableProperty(20))
-
-							weak var weakFirstProperty = firstProperty
-							weak var weakSecondProperty = secondProperty
-							weak var weakThirdProperty = thirdProperty
-
-							var outerProperty = Optional(MutableProperty(firstProperty!))
-							var flattened = Optional(outerProperty!.flatten(.Latest))
-
-							var errored = false
-							var completed = false
-
-							flattened!.producer.start { event in
-								switch event {
-								case .Completed:
-									completed = true
-								case .Failed:
-									errored = true
-								case .Interrupted, .Next:
-									break
-								}
-							}
-
-							firstProperty = nil
-							outerProperty!.value = secondProperty!
-							expect(weakFirstProperty).to(beNil())
-
-							secondProperty = nil
-							outerProperty!.value = thirdProperty!
-							expect(weakSecondProperty).to(beNil())
-
-							thirdProperty = nil
-							outerProperty = nil
-							flattened = nil
-							expect(weakThirdProperty).to(beNil())
-							expect(errored) == false
-							expect(completed) == true
-						}
-					}
-				}
-
-				describe("flatMap") {
-					describe("PropertyFlattenStrategy.Latest") {
-						it("should forward values from the latest inner transformed property") {
-							let firstProperty = Optional(MutableProperty(0))
-							var secondProperty = Optional(MutableProperty(10))
-							var thirdProperty = Optional(MutableProperty(20))
-
-							var outerProperty = Optional(MutableProperty(firstProperty!))
-
-							var receivedValues: [String] = []
-							var errored = false
-							var completed = false
-
-							outerProperty!.flatMap(.Latest) { $0.map { "\($0)" } }.producer.start { event in
-								switch event {
-								case let .Next(value):
-									receivedValues.append(value)
-								case .Completed:
-									completed = true
-								case .Failed:
-									errored = true
-								case .Interrupted:
-									break
-								}
-							}
-							
-							expect(receivedValues) == [ "0" ]
-							
-							outerProperty!.value = secondProperty!
-							secondProperty!.value = 11
-							outerProperty!.value = thirdProperty!
-							thirdProperty!.value = 21
-							
-							expect(receivedValues) == [ "0", "10", "11", "20", "21" ]
-							expect(errored) == false
-							expect(completed) == false
-							
-							secondProperty!.value = 12
-							secondProperty = nil
-							thirdProperty!.value = 22
-							thirdProperty = nil
-							
-							expect(receivedValues) == [ "0", "10", "11", "20", "21", "22" ]
-							expect(errored) == false
-							expect(completed) == false
-							
-							outerProperty = nil
-							expect(errored) == false
-							expect(completed) == true
-						}
-					}
-				}
-			}
-		}
-
-		describe("DynamicProperty") {
-			var object: ObservableObject!
-			var property: DynamicProperty<Int>!
-
-			let propertyValue: () -> Int? = {
-				if let value: AnyObject = property?.value {
-					return value as? Int
-				} else {
-					return nil
-				}
-			}
-
-			beforeEach {
-				object = ObservableObject()
-				expect(object.rac_value) == 0
-
-				property = DynamicProperty<Int>(object: object, keyPath: "rac_value")
-			}
-
-			afterEach {
-				object = nil
-			}
-
-			it("should read the underlying object") {
-				expect(propertyValue()) == 0
-
-				object.rac_value = 1
-				expect(propertyValue()) == 1
-			}
-
-			it("should write the underlying object") {
-				property.value = 1
-				expect(object.rac_value) == 1
-				expect(propertyValue()) == 1
-			}
-
-			it("should yield a producer that sends the current value and then the changes for the key path of the underlying object") {
-				var values: [Int] = []
-				property.producer.startWithNext { value in
-					expect(value).notTo(beNil())
-					values.append(value!)
-				}
-
-				expect(values) == [ 0 ]
-
-				property.value = 1
-				expect(values) == [ 0, 1 ]
-
-				object.rac_value = 2
-				expect(values) == [ 0, 1, 2 ]
-			}
-
-			it("should yield a producer that sends the current value and then the changes for the key path of the underlying object, even if the value actually remains unchanged") {
-				var values: [Int] = []
-				property.producer.startWithNext { value in
-					expect(value).notTo(beNil())
-					values.append(value!)
-				}
-
-				expect(values) == [ 0 ]
-
-				property.value = 0
-				expect(values) == [ 0, 0 ]
-
-				object.rac_value = 0
-				expect(values) == [ 0, 0, 0 ]
-			}
-
-			it("should yield a signal that emits subsequent values for the key path of the underlying object") {
-				var values: [Int] = []
-				property.signal.observeNext { value in
-					expect(value).notTo(beNil())
-					values.append(value!)
-				}
-
-				expect(values) == []
-
-				property.value = 1
-				expect(values) == [ 1 ]
-
-				object.rac_value = 2
-				expect(values) == [ 1, 2 ]
-			}
-
-			it("should yield a signal that emits subsequent values for the key path of the underlying object, even if the value actually remains unchanged") {
-				var values: [Int] = []
-				property.signal.observeNext { value in
-					expect(value).notTo(beNil())
-					values.append(value!)
-				}
-
-				expect(values) == []
-
-				property.value = 0
-				expect(values) == [ 0 ]
-
-				object.rac_value = 0
-				expect(values) == [ 0, 0 ]
-			}
-
-			it("should have a completed producer when the underlying object deallocates") {
-				var completed = false
-
-				property = {
-					// Use a closure so this object has a shorter lifetime.
-					let object = ObservableObject()
-					let property = DynamicProperty<Int>(object: object, keyPath: "rac_value")
-
-					property.producer.startWithCompleted {
-						completed = true
-					}
-
-					expect(completed) == false
-					expect(property.value).notTo(beNil())
-					return property
-				}()
-
-				expect(completed).toEventually(beTruthy())
-				expect(property.value).to(beNil())
-			}
-
-			it("should have a completed signal when the underlying object deallocates") {
-				var completed = false
-
-				property = {
-					// Use a closure so this object has a shorter lifetime.
-					let object = ObservableObject()
-					let property = DynamicProperty<Int>(object: object, keyPath: "rac_value")
-
-					property.signal.observeCompleted {
-						completed = true
-					}
-
-					expect(completed) == false
-					expect(property.value).notTo(beNil())
-					return property
-				}()
-
-				expect(completed).toEventually(beTruthy())
-				expect(property.value).to(beNil())
-			}
-
-			it("should retain property while DynamicProperty's underlying object is retained"){
-				weak var dynamicProperty: DynamicProperty<Int>? = property
-				
-				property = nil
-				expect(dynamicProperty).toNot(beNil())
-				
-				object = nil
-				expect(dynamicProperty).to(beNil())
-			}
-
-			it("should support un-bridged reference types") {
-				let dynamicProperty = DynamicProperty<UnbridgedObject>(object: object, keyPath: "rac_reference")
-				dynamicProperty.value = UnbridgedObject("foo")
-				expect(object.rac_reference.value) == "foo"
+		describe("map") {
+			it("should transform the current value and all subsequent values") {
+				let property = MutableProperty(1)
+				let mappedProperty = property
+					.map { $0 + 1 }
+					.map { $0 + 2 }
+
+				expect(mappedProperty.value) == 4
+
+				property.value = 2
+				expect(mappedProperty.value) == 5
 			}
 		}
 
@@ -1439,53 +679,6 @@ class PropertySpec: QuickSpec {
 					expect(bindingDisposable.disposed) == true
 				}
 			}
-
-			describe("to a dynamic property") {
-				var object: ObservableObject!
-				var property: DynamicProperty<Int>!
-
-				beforeEach {
-					object = ObservableObject()
-					expect(object.rac_value) == 0
-
-					property = DynamicProperty<Int>(object: object, keyPath: "rac_value")
-				}
-
-				afterEach {
-					object = nil
-				}
-
-				it("should bridge values sent on a signal to Objective-C") {
-					let (signal, observer) = Signal<Int, NoError>.pipe()
-					property <~ signal
-					observer.sendNext(1)
-					expect(object.rac_value) == 1
-				}
-
-				it("should bridge values sent on a signal producer to Objective-C") {
-					let producer = SignalProducer<Int, NoError>(value: 1)
-					property <~ producer
-					expect(object.rac_value) == 1
-				}
-
-				it("should bridge values from a source property to Objective-C") {
-					let source = MutableProperty(1)
-					property <~ source
-					expect(object.rac_value) == 1
-				}
-			}
 		}
-	}
-}
-
-private class ObservableObject: NSObject {
-	dynamic var rac_value: Int = 0
-	dynamic var rac_reference: UnbridgedObject = UnbridgedObject("")
-}
-
-private class UnbridgedObject: NSObject {
-	let value: String
-	init(_ value: String) {
-		self.value = value
 	}
 }
